# services/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CreatorHub API
  version: 1.0.0
  description: |
    CreatorHub 社区创作者平台 API。除健康检查、注册和登录接口外，接口均需要
    Authorization: Bearer <token> 请求头。
servers:
  - url: /
    description: 当前 API 服务
...
paths:
...
  /api/v1/short-dramas:
    get:
      tags: [ShortDramas]
      summary: 分页读取短剧
      operationId: listShortDramas
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 短剧列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortDramaListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    Unauthorized:
      description: 未登录或登录凭证无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseBase'
...
    ServerError:
      description: 服务端错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseBase'
  schemas:
    ResponseBase:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
          description: 业务数据；无数据的操作通常省略此字段
...
    ShortDramaEpisode:
      type: object
      required: [episode, title]
      properties:
        episode:
          type: integer
          minimum: 1
        title:
          type: string
        pwdId:
          type: string
          description: 夸克公开分享标识
        fid:
          type: string
        fidToken:
          type: string
        size:
          type: integer
          format: int64
        duration:
          type: integer
          format: int64
        url:
          type: string
          format: uri
          description: 最近一次采集到的临时地址，可能已过期
        m3u8url:
          type: string
          format: uri
    ShortDrama:
      type: object
      required: [id, sourceId, sourceName, externalId, name, totalEpisodes, episodes]
      properties:
        id:
          type: integer
          format: int64
        sourceId:
          type: integer
        sourceName:
          type: string
        externalId:
          type: string
        name:
          type: string
        updateTime:
          type: string
          format: date-time
        totalEpisodes:
          type: integer
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/ShortDramaEpisode'
    ShortDramaListResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ShortDrama'
...